+++ pyogrio/_compat.py
@@ -1,0 +1,28 @@
+from packaging.version import Version
+
+from pyogrio.core import __gdal_version__, __gdal_geos_version__
+
+# detect optional dependencies
+try:
+    import pyarrow
+except ImportError:
+    pyarrow = None
+
+try:
+    import shapely
+except ImportError:
+    shapely = None
+
+try:
+    import geopandas
+except ImportError:
+    geopandas = None
+
+
+HAS_ARROW_API = __gdal_version__ >= (3, 6, 0) and pyarrow is not None
+
+HAS_GEOPANDAS = geopandas is not None
+
+HAS_GDAL_GEOS = __gdal_geos_version__ is not None
+
+HAS_SHAPELY = shapely is not None and Version(shapely.__version__) >= Version("2.0.0")